Sketch out the route service
Our need is to have an app between the user and the statically served
file, that will be capable of retrieving some data from the request
being made and record that data in Google Analytics.

We're particularly interested in the format of the file the users are
accessing. There are two ways of obtaining that information:
- see path extension, ie. `resource.json`
- see `Accept` headers

In case of a scenario where both are provided, the path extension will
take priority.

The data is being used for further enhancing the product and potentially
dropping unused formats in the future.